+++ src/animation.c
@@ -1,14 +1,14 @@
 #include "animation.h"
 
-Animation anim_new(int steps) {
+Animation anim_new(size_t steps) {
 	return anim_new_al(al_new(sizeof(TextureRegion)), steps);
 }
 
-Animation anim_new_al(ArrayList frames, int steps) {
+Animation anim_new_al(ArrayList frames, size_t steps) {
 	return (Animation) { frames, 0, steps, 0 };
 }
 
-Animation anim_new_buffer(TextureRegion *frames, size_t num_frames, int steps) {
+Animation anim_new_buffer(TextureRegion *frames, size_t num_frames, size_t steps) {
 	return anim_new_al(al_prealloc(sizeof(TextureRegion), frames, num_frames), steps);
 }
